gcf/csv-migration/src/requirements.txt
pandas == 2.0.0
SQLAlchemy == 2.0.9
psycopg2-binary == 2.9.6
google-cloud-secret-manager == 2.16.1
google-cloud-storage == 2.8.0
